towards making ZonalDemandAggregator configurable
